Update patch set 1
Patch Set 1: Code-Review+2

Yes, those are Google specific tunings.
Also, there are a bunch of differences between this config and the common one in hw/qc/audio.
That one has broken speaker audio for me and earpiece audio is not actually coming out of the earpiece, but the speaker (just ... kind of quiet).

Patch-set: 1
Label: Code-Review=+2